+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- `match`: exported `ErrorMatcherFunc`
+
 
 ## [0.5.1] - 2020-01-06